Fix workspace_tools panic: INT4 column read as i64
The field_count column from jsonb_array_length returns INT4 (integer),
not INT8 (bigint). Reading it as Option<i64> causes a ColumnDecode
panic that kills the chat handler, resulting in immediate 502 errors
when users send messages.

+++ amos-harness/src/tools/workspace_tools.rs
@@ -61,7 +61,7 @@ impl Tool for GetWorkspaceSummaryTool {
                     json!({
                         "name": row.get::<String, _>("name"),
                         "display_name": row.get::<Option<String>, _>("display_name"),
-                        "field_count": row.get::<Option<i64>, _>("field_count").unwrap_or(0),
+                        "field_count": row.get::<Option<i32>, _>("field_count").unwrap_or(0),
                         "record_count": row.get::<Option<i64>, _>("record_count").unwrap_or(0),
                     })
                 })
